Finished renaming and scope
